backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pandas>=2.2.0
numpy>=1.26.3
pyarrow>=15.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
openai>=1.50.0
pydantic>=2.6.0
